# protocol/draft-2.0.0-1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HAIP — Human-Agent Interaction Protocol",
    "version": "2.0.0-draft.1",
    "description": "Draft portable review contract and execution extension. No HAIP 1 compatibility. Implementation ledger lists incomplete release gates.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
...
  "paths": {
    "/.well-known/haip": {
      "get": {
        "operationId": "get___well_known_haip",
        "summary": "Supported immutable revisions and profiles",
        "security": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Discovery"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or ambiguous input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Operation not permitted"
          },
          "404": {
            "description": "Resource not found or inaccessible"
          },
          "409": {
            "description": "State, ownership or idempotency conflict"
          },
          "410": {
            "description": "Private material deleted"
          },
          "413": {
            "description": "Uncompressed size limit"
          },
          "422": {
            "description": "Unsupported purpose/profile/renderer"
          },
          "429": {
            "description": "Captured quota exceeded"
          },
          "503": {
            "description": "Required authority, storage or identity unavailable"
          }
        }
      }
    },
    "/.well-known/haip-trust": {
      "get": {
        "operationId": "get___well_known_haip_trust",
        "summary": "Discovery only; never trust included keys automatically",
        "security": [],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrustManifest"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or ambiguous input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Operation not permitted"
          },
          "404": {
            "description": "Resource not found or inaccessible"
          },
          "409": {
            "description": "State, ownership or idempotency conflict"
          },
          "410": {
            "description": "Private material deleted"
          },
          "413": {
            "description": "Uncompressed size limit"
          },
          "422": {
            "description": "Unsupported purpose/profile/renderer"
          },
          "429": {
            "description": "Captured quota exceeded"
          },
          "503": {
            "description": "Required authority, storage or identity unavailable"
          }
        }
      }
    },
    "/v2/bundles": {
      "post": {
        "operationId": "post__v2_bundles",
        "summary": "Register a publisher-owned immutable app",
        "security": [
          {
            "producerBearer": []
          },
          {
            "humanSession": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 160
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewBundle"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or ambiguous input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Operation not permitted"
          },
          "404": {
            "description": "Resource not found or inaccessible"
          },
          "409": {
            "description": "State, ownership or idempotency conflict"
          },
          "410": {
            "description": "Private material deleted"
          },
          "413": {
            "description": "Uncompressed size limit"
          },
          "422": {
            "description": "Unsupported purpose/profile/renderer"
          },
          "429": {
            "description": "Captured quota exceeded"
          },
          "503": {
            "description": "Required authority, storage or identity unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BundleRegistration"
              }
            }
          }
        }
      }
    },
    "/v2/requests": {
      "post": {
        "operationId": "post__v2_requests",
        "summary": "Create one immutable review request",
        "security": [
          {
            "producerBearer": []
          },
          {
            "humanSession": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 160
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestStatus"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or ambiguous input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Operation not permitted"
          },
          "404": {
            "description": "Resource not found or inaccessible"
          },
          "409": {
            "description": "State, ownership or idempotency conflict"
          },
          "410": {
            "description": "Private material deleted"
          },
          "413": {
            "description": "Uncompressed size limit"
          },
          "422": {
            "description": "Unsupported purpose/profile/renderer"
          },
          "429": {
            "description": "Captured quota exceeded"
          },
          "503": {
            "description": "Required authority, storage or identity unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestInput"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "get__v2_requests",
        "summary": "List owned or assigned route requests",
        "security": [
          {
            "producerBearer": []
          },
          {
            "humanSession": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 99999999
            }
          },
          {
            "in": "query",
            "name": "state",
            "schema": {
              "type": "string",
              "enum": [
                "",
                "pending",
                "confirmed",
                "cancelled",
                "expired",
                "superseded"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestList"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or ambiguous input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Operation not permitted"
          },
          "404": {
            "description": "Resource not found or inaccessible"
          },
          "409": {
            "description": "State, ownership or idempotency conflict"
          },
          "410": {
            "description": "Private material deleted"
          },
          "413": {
            "description": "Uncompressed size limit"
          },
          "422": {
            "description": "Unsupported purpose/profile/renderer"
          },
          "429": {
            "description": "Captured quota exceeded"
          },
          "503": {
            "description": "Required authority, storage or identity unavailable"
          }
        }
      }
    },
    "/v2/requests/{id}": {
      "get": {
        "operationId": "get__v2_requests__id_",
        "summary": "Authoritative state; links are not credentials",
        "security": [
          {
            "producerBearer": []
          },
          {
            "humanSession": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestStatus"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or ambiguous input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Operation not permitted"
          },
          "404": {
            "description": "Resource not found or inaccessible"
          },
          "409": {
            "description": "State, ownership or idempotency conflict"
          },
          "410": {
            "description": "Private material deleted"
          },
          "413": {
            "description": "Uncompressed size limit"
          },
          "422": {
            "description": "Unsupported purpose/profile/renderer"
          },
          "429": {
            "description": "Captured quota exceeded"
          },
          "503": {
            "description": "Required authority, storage or identity unavailable"
          }
        }
      }
    },
    "/v2/requests/{id}/material": {
      "get": {
        "operationId": "get__v2_requests__id__material",
        "summary": "Retained payload, schema and canonical review document",
        "security": [
          {
            "producerBearer": []
          },
          {
            "humanSession": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Material"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or ambiguous input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Operation not permitted"
          },
          "404": {
            "description": "Resource not found or inaccessible"
          },
          "409": {
            "description": "State, ownership or idempotency conflict"
          },
          "410": {
            "description": "Private material deleted"
          },
          "413": {
            "description": "Uncompressed size limit"
          },
          "422": {
            "description": "Unsupported purpose/profile/renderer"
          },
          "429": {
            "description": "Captured quota exceeded"
          },
          "503": {
            "description": "Required authority, storage or identity unavailable"
          }
        }
      }
    },
    "/v2/requests/{id}/app": {
      "get": {
        "operationId": "get__v2_requests__id__app",
        "summary": "Stored app and input/result; authenticated human only",
        "security": [
          {
            "humanSession": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredApp"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or ambiguous input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Operation not permitted"
          },
          "404": {
            "description": "Resource not found or inaccessible"
          },
          "409": {
            "description": "State, ownership or idempotency conflict"
          },
          "410": {
            "description": "Private material deleted"
          },
          "413": {
            "description": "Uncompressed size limit"
          },
          "422": {
            "description": "Unsupported purpose/profile/renderer"
          },
          "429": {
            "description": "Captured quota exceeded"
          },
          "503": {
            "description": "Required authority, storage or identity unavailable"
          }
        }
      }
    },
    "/v2/requests/{id}/assignment": {
      "post": {
        "operationId": "post__v2_requests__id__assignment",
        "summary": "Renew reviewer assignment without authority",
        "security": [
          {
            "humanSession": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 160
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Assignment"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or ambiguous input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Operation not permitted"
          },
          "404": {
            "description": "Resource not found or inaccessible"
          },
          "409": {
            "description": "State, ownership or idempotency conflict"
          },
          "410": {
            "description": "Private material deleted"
          },
          "413": {
            "description": "Uncompressed size limit"
          },
          "422": {
            "description": "Unsupported purpose/profile/renderer"
          },
          "429": {
            "description": "Captured quota exceeded"
          },
          "503": {
            "description": "Required authority, storage or identity unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyInput"
              }
            }
          }
        }
      }
    },
    "/v2/requests/{id}/candidates": {
      "post": {
        "operationId": "post__v2_requests__id__candidates",
        "summary": "Freeze a schema-validated response",
        "security": [
          {
            "humanSession": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 160
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionCandidate"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or ambiguous input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Operation not permitted"
          },
          "404": {
            "description": "Resource not found or inaccessible"
          },
          "409": {
            "description": "State, ownership or idempotency conflict"
          },
          "410": {
            "description": "Private material deleted"
          },
          "413": {
            "description": "Uncompressed size limit"
          },
          "422": {
            "description": "Unsupported purpose/profile/renderer"
          },
          "429": {
            "description": "Captured quota exceeded"
          },
          "503": {
            "description": "Required authority, storage or identity unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionProposal"
              }
            }
          }
        }
      }
    },
    "/v2/requests/{id}/confirm": {
      "post": {
        "operationId": "post__v2_requests__id__confirm",
        "summary": "Confirm the exact frozen candidate as an authenticated human",
        "security": [
          {
            "humanSession": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 160
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Origin",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedRecord"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or ambiguous input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Operation not permitted"
          },
          "404": {
            "description": "Resource not found or inaccessible"
          },
          "409": {
            "description": "State, ownership or idempotency conflict"
          },
          "410": {
            "description": "Private material deleted"
          },
          "413": {
            "description": "Uncompressed size limit"
          },
          "422": {
            "description": "Unsupported purpose/profile/renderer"
          },
          "429": {
            "description": "Captured quota exceeded"
          },
          "503": {
            "description": "Required authority, storage or identity unavailable"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Confirmation"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Discovery": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "revisions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "profiles": {
            "$ref": "#/components/schemas/Profiles"
          },
          "renderer": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "ext_apps": {
                "type": "string"
              },
              "mcp_sdk": {
                "type": "string"
              }
            },
            "required": [
              "ext_apps",
              "mcp_sdk"
            ]
          },
          "mode": {
            "enum": [
              "development",
              "production"
            ]
          },
          "release_ready": {
            "type": "boolean"
          },
          "execution_admission": {
            "type": "string"
          },
          "notifications": {
            "enum": [
              "smtp",
              "polling-only"
            ]
          }
        },
        "required": [
          "name",
          "revisions",
          "profiles",
          "renderer",
          "mode",
          "release_ready",
          "execution_admission",
          "notifications"
        ]
      },
...
    },
    "securitySchemes": {
      "producerBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Separately provisioned machine credential; no human confirmation authority"
      },
      "humanSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-haip",
        "description": "OIDC server-side session; mutations also require CSRF and exact Origin"
      }
    }
  }
}